requirements.txt
mysql-connector-python
python-dotenv
requests
flask
flask-cors
pytest
pytest-html
xgboost
pandas
numpy
